interview-prep-assistant/requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
redis
celery
thefuzz
duckduckgo-search
python-dotenv
opentelemetry-api
opentelemetry-sdk
pytest
requests
groq
SQLAlchemy
psycopg2-binary
asyncpg
PyMuPDF
python-multipart
